app/src/test/java/mega/privacy/android/app/presentation/node/view/bottomsheetmenuitems/ShareBottomSheetMenuItemTest.kt
package mega.privacy.android.app.presentation.node.view.bottomsheetmenuitems

import kotlinx.coroutines.CoroutineScope
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.UnconfinedTestDispatcher
import kotlinx.coroutines.test.runTest
import mega.privacy.android.app.presentation.node.model.menuaction.ShareMenuAction
import mega.privacy.android.domain.entity.node.TypedFolderNode
import mega.privacy.android.domain.entity.node.TypedNode
import mega.privacy.android.domain.entity.shares.AccessPermission
import mega.privacy.android.domain.usecase.GetLocalFilePathUseCase
import mega.privacy.android.domain.usecase.file.GetFileByPathUseCase
import mega.privacy.android.domain.usecase.node.ExportNodeUseCase
import org.junit.jupiter.api.Assertions.*
import org.junit.jupiter.api.TestInstance
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.MethodSource
import org.mockito.kotlin.doReturn
import org.mockito.kotlin.mock
import java.util.stream.Stream

@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class ShareBottomSheetMenuItemTest {

    @OptIn(ExperimentalCoroutinesApi::class)
    private val applicationScope: CoroutineScope = CoroutineScope(
        UnconfinedTestDispatcher()
    )

    private val getLocalFilePathUseCase: GetLocalFilePathUseCase = mock()
    private val exportNodesUseCase: ExportNodeUseCase = mock()
    private val getFileByPathUseCase: GetFileByPathUseCase = mock()
    private val shareBottomSheetMenuItem = ShareBottomSheetMenuItem(
        menuAction = ShareMenuAction(),
        scope = applicationScope,
        getLocalFilePathUseCase = getLocalFilePathUseCase,
        exportNodesUseCase = exportNodesUseCase,
        getFileByPathUseCase = getFileByPathUseCase
    )

    @ParameterizedTest(name = "isNodeInRubbish {0} - accessPermission {1} - isInBackups {2} - node {3} - isConnected {4} - expected {5}")
    @MethodSource("provideTestParameters")
    fun `test that share bottom sheet menu item visibility is correct`(
        isNodeInRubbish: Boolean,
        accessPermission: AccessPermission?,
        isInBackups: Boolean,
        node: TypedNode,
        isConnected: Boolean,
        expected: Boolean,
    ) = runTest {
        val result = shareBottomSheetMenuItem.shouldDisplay(
            isNodeInRubbish,
            accessPermission,
            isInBackups,
            node,
            isConnected,
        )
        assertEquals(expected, result)
    }

    private fun provideTestParameters() = Stream.of(
        Arguments.of(
            true,
            AccessPermission.OWNER,
            false,
            mock<TypedFolderNode> { on { isTakenDown } doReturn true },
            false,
            false,
        ),
        Arguments.of(
            true,
            AccessPermission.FULL,
            false,
            mock<TypedFolderNode> { on { isTakenDown } doReturn false },
            false,
            false,
        ),
        Arguments.of(
            false,
            AccessPermission.OWNER,
            false,
            mock<TypedFolderNode> { on { isTakenDown } doReturn false },
            false,
            true,
        )
    )
}